최종 2차 코드리뷰

+++ catalog-service/src/main/java/com/node5/catalogservice/search/infrastructure/elasticsearch/ElasticsearchProductSearchAdapter.java
@@ -42,11 +42,11 @@ public class ElasticsearchProductSearchAdapter implements ProductSearchPort {
 
 			// filter:
 			// - 판매 중인 상품만 자동완성 대상
-			// - 점수 계산과 무관한 고정 조건
-			b.filter(f -> f.term(t -> t.field("status").value(FieldValue.of("ON_SALE"))));
+			b.filter(f -> f.term(t -> t.field("status")
+				.value(FieldValue.of("ON_SALE"))));
 
 			// must:
-			// - 자동완성 전용 필드(name_autocomplete)로 prefix 매칭
+			// - 자동완성 필드(name_autocomplete)로 검색어 매칭
 			if (StringUtils.hasText(command.keyword())) {
 				b.must(m -> m.match(mm -> mm
 					.field("name_autocomplete")
@@ -63,8 +63,7 @@ public class ElasticsearchProductSearchAdapter implements ProductSearchPort {
 			.withMaxResults(AUTOCOMPLETE_LIMIT)
 			.build();
 
-		SearchHits<ProductDocument> hits =
-			elasticsearchOperations.search(nativeQuery, ProductDocument.class);
+		SearchHits<ProductDocument> hits = elasticsearchOperations.search(nativeQuery, ProductDocument.class);
 
 		return hits.getSearchHits().stream()
 			.map(hit -> hit.getContent().getName())
@@ -76,8 +75,7 @@ public class ElasticsearchProductSearchAdapter implements ProductSearchPort {
 	public Page<ProductDocument> search(ProductSearchCommand command, Pageable pageable) {
 		Query query = buildBoolQuery(command);
 
-		ProductSearchSort sort =
-			command.sort() != null ? command.sort() : ProductSearchSort.LATEST;
+		ProductSearchSort sort = command.sort() != null ? command.sort() : ProductSearchSort.LATEST;
 
 		NativeQuery nativeQuery = new NativeQueryBuilder()
 			.withQuery(query)
@@ -85,11 +83,9 @@ public class ElasticsearchProductSearchAdapter implements ProductSearchPort {
 			.withSort(buildSortOptions(sort))
 			.build();
 
-		SearchHits<ProductDocument> hits =
-			elasticsearchOperations.search(nativeQuery, ProductDocument.class);
+		SearchHits<ProductDocument> hits = elasticsearchOperations.search(nativeQuery, ProductDocument.class);
 
-		SearchPage<ProductDocument> page =
-			SearchHitSupport.searchPageFor(hits, pageable);
+		SearchPage<ProductDocument> page = SearchHitSupport.searchPageFor(hits, pageable);
 
 		return page.map(hit -> hit.getContent());
 	}
@@ -98,28 +94,28 @@ public class ElasticsearchProductSearchAdapter implements ProductSearchPort {
 		return Query.of(q -> q.bool(b -> {
 
 			// filter:
-			// - 점수(scoring)에 영향을 주지 않는 조건
-			// - 캐시 가능 → 성능 유리
-			// - 항상 적용되는 필터 조건에 적합
-			b.filter(f -> f.term(t -> t.field("status").value(FieldValue.of("ON_SALE"))));
+			// - 판매 중인 상품만 검색 대상
+			// - 결과를 걸러내는 조건 (점수에는 영향 없음)
+			b.filter(f -> f.term(t -> t.field("status")
+				.value(FieldValue.of("ON_SALE"))));
 
 			// filter:
-			// - exact match 조건
-			// - 검색 결과 점수에 영향 없음
+			// - 판매자 기준으로 결과 제한
 			if (command.shopId() != null) {
-				b.filter(f -> f.term(t -> t.field("shopId").value(FieldValue.of(command.shopId().toString()))));
+				b.filter(f -> f.term(t -> t.field("shopId")
+					.value(FieldValue.of(command.shopId().toString()))));
 			}
 
 			// filter:
-			// - enum 값 정확 일치 조건
+			// - 카테고리 기준으로 결과 제한
 			if (command.category() != null) {
-				b.filter(f -> f.term(t -> t.field("category").value(FieldValue.of(command.category().name()))));
+				b.filter(f -> f.term(t -> t.field("category")
+					.value(FieldValue.of(command.category().name()))));
 			}
 
 			// must:
-			// - 텍스트 검색 조건
-			// - relevance score 계산 대상
-			// - 키워드 검색은 점수 기반 정렬에 의미가 있으므로 must 사용
+			// - 검색어 매칭
+			// - 여기서 점수 계산됨 (결과 순서에 영향)
 			if (StringUtils.hasText(command.keyword())) {
 				b.must(m -> m.match(mm -> mm
 					.field("name")
@@ -129,17 +125,17 @@ public class ElasticsearchProductSearchAdapter implements ProductSearchPort {
 			}
 
 			// filter:
-			// - 범위 조건은 점수와 무관
-			// - 숫자 필터는 filter로 처리하여 성능 최적화
+			// - 가격 범위로 결과 제한
+			// - 점수에는 영향 없음
 			if (command.minPrice() != null && command.maxPrice() != null) {
-				long min = command.minPrice().longValue();
-				long max = command.maxPrice().longValue();
+				double min = command.minPrice().doubleValue();
+				double max = command.maxPrice().doubleValue();
 
 				b.filter(f -> f.range(r -> r
 					.number(n -> n
 						.field("price")
-						.gte((double) min)
-						.lte((double) max)
+						.gte(min)
+						.lte(max)
 					)
 				));
 			}
